Fixed warnings

+++ include/RSAEngine.h
@@ -71,7 +71,7 @@ class RSANumber {
             *this = 0;
             int index = 0;
 
-            for(int i = binary.size() - 1; i >= 0; --i) {
+            for(size_t i = binary.size() - 1; i >= 0; --i) {
                 if(binary[i] == '1') {
                     setBit(index);
                 }
